combined and deleted individual api files to one main API
