--- client/src/utils/EventAPI.js
@@ -1,20 +1,0 @@
-import axios from "axios";
-
-export default {
-  // Gets all events
-  getEvents: function() {
-    return axios.get("/api/events");
-  },
-  // Gets the book with the given id
-  getEvent: function(id) {
-    return axios.get("/api/events/" + id);
-  },
-  // Deletes the book with the given id
-  deleteEvent: function(id) {
-    return axios.delete("/api/events/" + id);
-  },
-  // Saves a book to the database
-  saveEvent: function(eventData) {
-    return axios.post("/api/events", eventData);
-  }
-};
