--- client/src/utils/UsersAPI.js
@@ -1,20 +1,0 @@
-import axios from "axios";
-
-export default {
-  // Gets all users
-  getUsers: function() {
-    return axios.get("/api/users");
-  },
-  // Gets the book with the given id
-  getUser: function(id) {
-    return axios.get("/api/users/" + id);
-  },
-  // Deletes the book with the given id
-  deleteUser: function(id) {
-    return axios.delete("/api/users/" + id);
-  },
-  // Saves a book to the database
-  saveUser: function(userData) {
-    return axios.post("/api/users", userData);
-  }
-};
