--- client/src/utils/VendorAPI.js
@@ -1,20 +1,0 @@
-import axios from "axios";
-
-export default {
-  // Gets all vendors
-  getVendors: function() {
-    return axios.get("/api/vendors");
-  },
-  // Gets the book with the given id
-  getVendor: function(id) {
-    return axios.get("/api/vendors/" + id);
-  },
-  // Deletes the book with the given id
-  deleteVendor: function(id) {
-    return axios.delete("/api/vendors/" + id);
-  },
-  // Saves a book to the database
-  saveVendor: function(vendorData) {
-    return axios.post("/api/vendors", vendorData);
-  }
-};
