+++ routes/api/index.js
@@ -3,7 +3,7 @@ const vendorRoutes = require("./vendor-routes");
 const eventRoutes = require("./event-routes");
 const usersRoutes = require("./users-routes");
 
-// routes
+// Table routes
 router.use("/vendors", vendorRoutes);
 router.use("/events", eventRoutes);
 router.use("/users", usersRoutes);
